+++ generated_cpp_bindings/x3d/nodes/CADPart.hpp
@@ -153,9 +153,9 @@ public:
   /**
    * @brief Gets the value of children. AccessType: inputOutput
    * @details Grouping nodes contain an ordered list of children nodes.
-   * @return MFNode The current value of children.
+   * @return const MFNode& The current value of children.
    */
-  MFNode getChildren() const { return _children; }
+  const MFNode &getChildren() const { return _children; }
 
   /**
    * @brief Acceptable node types for the children field.