+++ generated_cpp_bindings/x3d/nodes/ComposedShader.hpp
@@ -81,9 +81,9 @@ public:
    * @brief Gets the value of field. AccessType: inputOutput
    * @details Include a field statement for each field declaration in the
    * ComposedShader node.
-   * @return MFNode The current value of field.
+   * @return const MFNode& The current value of field.
    */
-  MFNode getField() const { return _field; }
+  const MFNode &getField() const { return _field; }
 
   /**
    * @brief Acceptable node types for the field field.
@@ -165,9 +165,9 @@ public:
    * @brief Gets the value of parts. AccessType: inputOutput
    * @details ComposedShader can contain multiple ShaderPart nodes in the parts
    * field.
-   * @return MFNode The current value of parts.
+   * @return const MFNode& The current value of parts.
    */
-  MFNode getParts() const { return _parts; }
+  const MFNode &getParts() const { return _parts; }
 
   /**
    * @brief Acceptable node types for the parts field.