+++ generated_cpp_bindings/x3d/nodes/ComposedTexture3D.hpp
@@ -68,9 +68,9 @@ public:
   /**
    * @brief Gets the value of texture. AccessType: inputOutput
    * @details collection of 2D texture sources.
-   * @return MFNode The current value of texture.
+   * @return const MFNode& The current value of texture.
    */
-  MFNode getTexture() const { return _texture; }
+  const MFNode &getTexture() const { return _texture; }
 
   /**
    * @brief Acceptable node types for the texture field.