+++ generated_cpp_bindings/x3d/nodes/ViewpointGroup.hpp
@@ -107,9 +107,9 @@ public:
    * @details ViewpointGroup contains Viewpoint, OrthoViewpoint, GeoViewpoint
    * and other ViewpointGroup nodes that each have containerField='children'
    * default value.
-   * @return MFNode The current value of children.
+   * @return const MFNode& The current value of children.
    */
-  MFNode getChildren() const { return _children; }
+  const MFNode &getChildren() const { return _children; }
 
   /**
    * @brief Acceptable node types for the children field.